test/requirements.txt
python-telegram-bot==22.5
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.32.5
urllib3==2.5.0